requirements.txt
pandas
python-dotenv
MetaTrader5
plotly
numpy